Split out IntegrationTests into separate testTarget
