+++ Tests/PostgresNIOTests/New/Extensions/LoggingUtils.swift
@@ -1,9 +1,2 @@
 import Logging
 
-extension Logger {
-    static var psqlTest: Logger {
-        var logger = Logger(label: "psql.test")
-        logger.logLevel = .info
-        return logger
-    }
-}
